+++ src/io/hdf5/vec_klgs_2010_io.cpp
@@ -19,7 +19,7 @@ VecKlgs2010Input read_input(const ModelFile &file)
     read_mat_if_present(file, "/data/train/y", input.train.y);
     read_mat_if_present(file, "/data/train/w", input.train.w);
     read_mat_if_present(file, "/data/train/x", input.train.x);
-    read_mat_if_present(file, "/data/forecast/z", input.forecast.z);
+    read_forecast_regressors(file, input.spec.k, input.forecast.x);
 
     if (input.use_a())
     {