+++ src/io/hdf5/vec_normal_stochvol_io.cpp
@@ -52,7 +52,7 @@ VecNormalStochvolInput read_input(const ModelFile &file)
     read_mat_if_present(file, "/data/train/y", input.train.y);
     read_mat_if_present(file, "/data/train/w", input.train.w);
     read_mat_if_present(file, "/data/train/z", input.train.z);
-    read_mat_if_present(file, "/data/forecast/z", input.forecast.z);
+    read_forecast_regressors(file, input.spec.k, input.forecast.x);
 
     if (input.use_a())
     {